# Dynamic shape latency Step1 : Cache compile-time information in FusionExecutors