--- a/torch/csrc/jit/codegen/cuda/executor.h
+++ b/torch/csrc/jit/codegen/cuda/executor.h
@@ -81,6 +81,9 @@ class TORCH_CUDA_CU_API FusionExecutor : public NonCopyable {
     uint64_t rand_offset;
   };
 
+  using ExecutorCompileTimeInfoCache =
+      executor_utils::caching::ExecutorCompileTimeInfoCache;
+
   kir::Kernel* kernel() const {
     return lowered_.kernel();
   }
@@ -169,6 +172,10 @@ class TORCH_CUDA_CU_API FusionExecutor : public NonCopyable {
     return used_tvs_;
   };
 
+  ExecutorCompileTimeInfoCache* compileTimeDataCache() {
+    return &compile_time_info_cache_;
+  }
+
  private:
   Fusion fusion_;
 
@@ -201,6 +208,9 @@ class TORCH_CUDA_CU_API FusionExecutor : public NonCopyable {
 
   // Profiling support: knob to disable caching of launch params
   bool disable_parameter_cache_ = false;
+
+  // Compile time information caching
+  ExecutorCompileTimeInfoCache compile_time_info_cache_;
 };
 
 } // namespace cuda