… benchmarks/cpp/nvfuser/shape_inference.cpp
… torch/csrc/jit/codegen/cuda/executor_utils.h