Dynamic shape runtime latency mainly includes two parts:
1. Searching the `FusionExecutorCache` for compiled kernels
2. Run shape inference on input to derive launch bounds and output allocations. 

This PR speeds up Part 2, by caching the compile-time available information in the `FusionExecutor`'s so that they do not have to be re-computed at runtime. 


Perf comparison:

```
Before:
@ 9a1910add0c726f9f403d2444b321d2f30dc386b
-------------------------------------------------------------------------------------------
Benchmark                                                 Time             CPU   Iterations
-------------------------------------------------------------------------------------------
LayerNormBackward_ShapeInference                       87.1 us         87.1 us         8171
LayerNormForward_ShapeInference                        42.9 us         42.9 us        15433
LayerNormBackward_NoShapeInferenceCachedBaseline       13.0 us         13.0 us        53089
LayerNormForward_NoShapeInferenceCachedBaseline        4.56 us         4.56 us       155398

After:
@ 96a2353b30e8874da80228407ff0ad92425d7072
-------------------------------------------------------------------------------------------
Benchmark                                                 Time             CPU   Iterations
-------------------------------------------------------------------------------------------
LayerNormBackward_ShapeInference                       41.7 us         41.7 us        15225
LayerNormForward_ShapeInference                        22.7 us         22.7 us        28692
LayerNormBackward_NoShapeInferenceCachedBaseline       12.0 us         12.0 us        56930
LayerNormForward_NoShapeInferenceCachedBaseline        4.38 us         4.38 us       162971
```